Add Audio Test Script
The script has examples of how to use FMOD and the different ways to approach implementing a sound when doing so. There is the method of Scriptable Object which allows for a quick access of the full library of audio from FMOD and then there is the Event Ref which allows for an inspector approach.